@@ -5,12 +5,12 @@
 
     <groupId>ca.qc.inspq</groupId>
     <artifactId>utilitaire-nam</artifactId>
-    <version>tag1112-SNAPSHOT</version>
+    <version>0.2.0-SNAPSHOT</version>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>1.3.5.RELEASE</version>
+        <version>2.2.0.RELEASE</version>
     </parent>
 
     <packaging>pom</packaging>
@@ -46,17 +46,25 @@
 
     <properties>
         <commons-lang3.version>3.4</commons-lang3.version>
-        <java.version>1.7</java.version>
+        <java.version>11</java.version>
         <javax.ws.rs.version>2.0</javax.ws.rs.version>
         <maven-compiler-plugin.version>3.3</maven-compiler-plugin.version>
         <resteasy.version>3.0.13.Final</resteasy.version>
         <resteasy-spring-boot-starter.version>2.0.0-RELEASE</resteasy-spring-boot-starter.version>
-        <vaadin.version>7.6.6</vaadin.version>
-        <vaadin-spring-boot-starter.version>1.0.0</vaadin-spring-boot-starter.version>
+        <vaadin.version>8.8.0</vaadin.version>
+        <vaadin-spring-boot-starter.version>3.2.1</vaadin-spring-boot-starter.version>
     </properties>
 
     <modules>
         <module>utilitaire-NAM-API</module>
         <module>utilitaire-NAM-Service</module>
     </modules>
+    
+    <dependencies>
+    	<dependency>
+		    <groupId>org.springframework.boot</groupId>
+		    <artifactId>spring-boot-starter-test</artifactId>
+		    <scope>test</scope>
+		</dependency>
+    </dependencies>
 </project>
