@@ -2,10 +2,10 @@ package ca.qc.inspq.nam;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.web.SpringBootServletInitializer;
 
 @SpringBootApplication
-public class Application extends SpringBootServletInitializer {
+public class Application {
+	
     public static void main(String[] args) {
         SpringApplication.run(Application.class, args);
     }
